% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{fdr_fun}
\alias{fdr_fun}
\title{Calculates FDR when ground truth is available}
\usage{
fdr_fun(z, id)
}
\arguments{
\item{z}{Vector of cluster assignments}

\item{id}{Vector of true cluster assignments (ground truth)}
}
\value{
FDR
}
\description{
Calculates FDR when ground truth is available
}
