% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{mean_fdr}
\alias{mean_fdr}
\title{Calculates average FDR when ground truth is available}
\usage{
mean_fdr(zm, id)
}
\arguments{
\item{zm}{Matrix with posterior samples of cluster assignments}

\item{id}{Vector of true cluster assignments (ground truth)}
}
\value{
Average FDR over posterior samples
}
\description{
Calculates average FDR when ground truth is available
}
