% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{fnr_fun}
\alias{fnr_fun}
\alias{mean_fnr}
\title{Calculates false negative rate (FNR) when the ground truth is available}
\usage{
fnr_fun(z, id, use_apply = F)

mean_fnr(zm, id, use_apply = F)
}
\arguments{
\item{z}{Integer vector of cluster assignments}

\item{id}{Integer vector of true cluster assignments (ground truth)}

\item{use_apply}{Logical (default F), whether to use \code{apply()} to calculate the rate. 
Setting \code{use_apply = T} may be slower but memory efficient when \code{length(z)} is large.}

\item{zm}{Matrix with posterior samples of cluster assignments, where each row corresponds to one sample from the posterior}
}
\value{
False negative rate (FNR)
}
\description{
\loadmathjax False negative rate (FNR) of the estimated record linkage (partition) based on the ground truth is defined as (Steorts, 2015)
\mjsdeqn{FNR = \frac{FN}{CL + FN}}
where FN is the number of false negatives (linked under the truth but not linked under the estimate) and CL is the number of correct links (true positives). 
If both FN=0 and CL=0, define FNR = 0. 

FNR can be also defined as \mjseqn{FNR = 1 - Recall}, where \mjseqn{Recall = CL/(CL+FN)}.

\code{fnr_fun} calculates FNR for an estimated partition, and \code{mean_fnr} calculates average FNR based on posterior samples of partition.
}
\examples{
nclusters_per_size <- c(50,50,50,50)
numberFields <- 5
numberCategories <- rep(10,5)
trueBeta <- 0.01
# generate simulated data
simulatedData <- SimData(nclusters_per_size, nfields = numberFields, ncat = numberCategories, true_beta = trueBeta)
# Fit ESCD model
posteriorESCD <- SampleCluster(data=simulatedData, Prior="ESCD", burn=5, nsamples=10)
# true number of clusters
trueK = sum(nclusters_per_size)
# true cluster membership vector
trueid = rep(1:trueK, times=rep(1:length(nclusters_per_size), times=nclusters_per_size))
# FNR calculation for a single estimate
fnr_fun(posteriorESCD$Z[1,], trueid)
# average FNR calculation
mean_fnr(posteriorESCD$Z, trueid)

}
\references{
Steorts, R. C. (2015). Entity resolution with empirically motivated priors. Bayesian Analysis, 10(4), 849-875.
}
\seealso{
\code{\link{fdr_fun}}, \code{\link{mean_fdr}}
}
