% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{fnr_fun}
\alias{fnr_fun}
\title{Calculates FNR when ground truth is available}
\usage{
fnr_fun(z, id)
}
\arguments{
\item{z}{Vector of cluster assignments}

\item{id}{Vector of true cluster assignments (ground truth)}
}
\value{
FNR
}
\description{
Calculates FNR when ground truth is available
}
