% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{mean_fnr}
\alias{mean_fnr}
\title{Calculates average FNR when ground truth is available}
\usage{
mean_fnr(zm, id)
}
\arguments{
\item{zm}{Matrix with posterior samples of cluster assignments}

\item{id}{Vector of true cluster assignments (ground truth)}
}
\value{
Average FNR over posterior samples
}
\description{
Calculates average FNR when ground truth is available
}
