% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesESC.R
\name{SampleCluster}
\alias{SampleCluster}
\title{Posterior samples of cluster assignments and Prior parameters}
\usage{
SampleCluster(data, Prior, burn, nsamples, spacing = 1000, block_flag = TRUE)
}
\arguments{
\item{data}{Data frame containing only categorical variables}

\item{Prior}{Specify partition prior: "DP", "PY", "ESCNB"}

\item{burn}{MCMC burn-in period}

\item{nsamples}{MCMC iterations after burn-in}

\item{spacing}{Thinning for chaperones algorithm (default 1000)}

\item{block_flag}{TRUE for non-uniform chaperones (default)}
}
\value{
List with posterior samples for cluster assignments (Z),
        Prior parameters and distortion probabilities (Params)
}
\description{
Posterior samples of cluster assignments and Prior parameters
}
