% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesESC.R
\name{SampleCluster}
\alias{SampleCluster}
\title{Fitting Bayesian entity resolution models for categorical data}
\usage{
SampleCluster(
  data,
  Prior,
  burn,
  nsamples,
  spacing = 1000,
  block_flag = TRUE,
  beta_known = NULL,
  Nbinom = NULL
)
}
\arguments{
\item{data}{Integer matrix, noisy categorical data where row corresponds to records and column corresponds to fields.}

\item{Prior}{Specify partition prior: "DP", "PY", "ESCNB", "ESCD", "ESCP", "ESCB", "ESCBshift". See \strong{Details} section for their properties and differences.}

\item{burn}{MCMC burn-in period}

\item{nsamples}{MCMC iterations after burn-in}

\item{spacing}{Thinning for chaperones algorithm (default 1000)}

\item{block_flag}{TRUE for non-uniform chaperones (default)}

\item{beta_known}{default NULL, specify if true distortion probabilities (with length same as \code{ncol(data)}) is assumed to be known.}

\item{Nbinom}{default NULL, specify if Prior is "ESCB" which represents the maximum}
}
\value{
List with posterior samples for cluster assignments (Z),
Prior parameters and distortion probabilities (Params)
}
\description{
\code{SampleCluster} runs Bayesian entity resolution model for categorical dataset under various possible random partition priors.
The possible choices of random partition prior for the \code{Prior} argument are:
\itemize{
\item \code{"DP"}: Random partition induced by Dirichlet process (DP), also known as Chinese restaurant process prior.
\item \code{"PY"}: Random partition induced by Pitman-Yor process (PY).
\item \code{"ESCNB"}: Exchangeable sequence of clusters (ESC) model with zero-truncated negative binomial distribution.
\item \code{"ESCD"}: ESC model with Dirichlet distribution.
\item \code{"ESCP"}: ESC model with zero-truncated Poisson distribution.
\item \code{"ESCB"}: ESC model with zero-truncated binomial distribution with fixed maximum cluster size (given by \code{Nbinom}).
\item \code{"ESCBshift"}: ESC model with shifted binomial distribution, non-fixed maximum cluster size.
}
The \strong{Details} section below describes the properties of these priors. Especially for the details of ESC models, see Betancourt, Zanella and Steorts (2020) and Lee and Sang (2022).

The independent fields model of Steorts, Hall, and Fienberg (2016) has been used to model the noisy categorical data.
Specifically, the density vector \eqn{\theta_l} within each field is fixed as the empirical distribution of the data; See Section 4 of Betancourt et al. (2020) for detailed specification.
}
\details{
The choice of random partition prior significantly affects the Bayesian entity resolution results.
There are two major properties of random partition priors that affects the entity resolution result: \strong{microclustering} and \strong{balancedness}.

\strong{Microclustering.} Traditional exchangeable random partition models such as the one induced by Dirichlet process (DP) or Pitman-Yor process (PY), assumes that
the number of data points in each cluster grows linearly with the total number of data points.
This growth assumption is not desirable in entity resolution tasks, where the cluster represents duplicates of the record so that
cluster sizes remain small, even for large data sets.
The \emph{microclustering} property (Zanella, Betancourt and others, 2016) holds when cluster sizes grow sublinearly with the total number of data points,
and the exchangeable sequence of clusters (ESC) model (Betancourt et al, 2020) is a very general class of random partition models that possess microclustering property.
ESC models can directly control the prior distribution of the cluster sizes (\eqn{\mu}), which can be either zero-truncated negative binomial distribution, zero-truncated Poisson, and many others.

\strong{Balancedness.} Traditional exchangeable random partition models such as Chinese restaurant process induced by DP, often possesses the "rich-get-richer" property.
This gives tendency that some few clusters become large and dominates the whole dataset a priori, leading to an unbalanced partition.
Lee and Sang (2022) studied the balancedness of random partition models, and proposed \emph{balance-averse} and \emph{balance-seeking} properties
which corresponds to favoring unbalanced and balanced partition in terms of probability, respectively.
While the microclustering property has a similiar rationale by limiting the growth rate of the largest cluster,
the balancedness property analyzes how it assigns probabilities to partitions with different levels of balancedness in non-asymptotic point of view and they complement each other.

The table below summarizes the properties with different choice of random partition priors:\tabular{lll}{
   \code{Prior} \tab Microclustering \tab Balancedness \cr
   \code{"DP"} \tab No \tab balance-averse \cr
   \code{"PY"} \tab No \tab balance-averse \cr
   \code{"ESCNB"} \tab Yes \tab balance-averse \cr
   \code{"ESCD"} \tab Yes \tab N/A \cr
   \code{"ESCP"} \tab Yes \tab balance-neutral \cr
   \code{"ESCB"} \tab Yes, bounded microclustering \tab balance-seeking \cr
   \code{"ESCBshift"} \tab Yes \tab balance-seeking \cr
}


Here \code{Prior = "ESCD"} is neither balance-averse nor balance-seeking,
and \code{Prior = "ESCB"} has \emph{bounded microclustering propoerty} (Betancourt et al, 2022), where the maximum size of the cluster is upper bounded by the fixed hyperparameter \code{Nbinom}.
To let the binomial parameter (number of trials) also be random, use \code{Prior = "ESCBShift"}.
Using balance-seeking prior leads to assigning less prior probability to partition with many singleton clusters, thus reguarlizing the emergence of singleton clusters.

For the posterior sampling of cluster assignments (partition), a tailored MCMC scheme named \emph{chaperones algorithm} (Zanella, Betancourt and others, 2016) has been used.
}
\examples{
nclusters_per_size <- c(50,50,50,50)
numberFields <- 5
numberCategories <- rep(10,5)
trueBeta <- 0.01
# generate simulated data
simulatedData <- SimData(nclusters_per_size, numberFields, numberCategories, trueBeta)
# Fit ESCD model
posteriorESCD <- SampleCluster(data=simulatedData, Prior="ESCD", burn=0, nsamples=10)

}
\references{
Steorts, R. C., Hall, R., & Fienberg, S. E. (2016). A Bayesian approach to graphical record linkage and deduplication. Journal of the American Statistical Association, 111(516), 1660-1672.

Zanella, G., Betancourt, B., Miller, J. W., Wallach, H., Zaidi, A., & Steorts, R. C. (2016). Flexible models for microclustering with application to entity resolution. Advances in neural information processing systems, 29.

Betancourt, B., Zanella, G., & Steorts, R. C. (2020). Random partition models for microclustering tasks. Journal of the American Statistical Association, 1-13.

Betancourt, B., Sosa, J., & Rodríguez, A. (2022). A prior for record linkage based on allelic partitions. Computational Statistics & Data Analysis, 172, 107474.

Lee, C. J., & Sang, H. (2022). Proceedings of the 39th International Conference on Machine Learning (ICML), PMLR 162:12521 - 12541.
}
