% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{SimData}
\alias{SimData}
\title{Generates a simulated dataset based on a true partition}
\usage{
SimData(true_L, nfields, ncat, true_beta)
}
\arguments{
\item{true_L}{Vector of size max cluster size with number of clusters of each size}

\item{nfields}{Number of fields}

\item{ncat}{Vector with number of categories per field}

\item{true_beta}{Distortion probability for the fields}
}
\value{
Simulated data set
}
\description{
Generates a simulated dataset based on a true partition
}
